impact-of-disruptions-on-firms/requirements.txt
pandas
unidecode
requests
python-dotenv
folium